fix: map C# byte→Java int, sbyte→Java byte, with &0xFF masking
Systematic fix across the entire conversion pipeline:
- TypeMappings.json + TypeMappingService: byte→int mapping root
- ExpressionTransformerHelpers: &0xFF masking for byte narrowing, wrapper type fix
- IdentifierExpressionTransformer: boxed class name + PredefinedType
- TypeOperationTransformer: default(byte)→0, sizeof(byte)→4
- AssignmentTransformer: compound assignment &0xFF for byte targets
- ArgumentTransformer: C# byte→Java byte param narrowing cast
- InvocationExpressionTransformer: wrapper type for primitive special type
- byte[] arrays stay Java byte[] for API compatibility

+++ src/CSharpToJava.Core/Transformers/Expression/Transformers/InvocationExpressionTransformer.cs
@@ -5077,7 +5077,7 @@ public class InvocationExpressionTransformer : IIRExpressionTransformer
             SpecialType.System_Int32   => "Integer",
             SpecialType.System_Int64   => "Long",
             SpecialType.System_Int16   => "Short",
-            SpecialType.System_Byte    => "Byte",
+            SpecialType.System_Byte    => "Integer",
             SpecialType.System_SByte   => "Byte",
             SpecialType.System_UInt32  => "Integer",
             SpecialType.System_UInt64  => "Long",
